requirements.txt
spotipy
python-dotenv
fastapi
uvicorn[standard]
